WIP: convert to postgres using asyncpg
This should improve the bot's performance.

Note that this code is untested and is a minimum effort to make things
work. The poll handling functions are a mess and contain many bits of
repeated code, should be refactored later.

I've also removed the database initialization from the entry script for
now, so it has to be done manually by applying `init.sql` to the
database.

Also note that postgresql is not optional now, since maintaining support
for multiple sql backend is a pain.